# chore: Use boltdb Batch API for metastore writes