**What this PR does / why we need it**:
* Boltdb performs quite slowly when using `Update` as every call will result in a flush to disk
* Switching to batch allows calls from multiple goroutines (API calls) to be batched together into a single flush to disk which increase performance significantly.
* Default parameters for batching are max 1000 pending writes or every 10ms so each write still returns within a reasonable time.
* The benchmark scales roughly linearly up to 1000 workers: ~150 ops/sec to >80,000 ops/sec

```
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/v3/pkg/ingester-rf1/metastore
                                 │  before.txt   │              after.txt              │
                                 │    sec/op     │   sec/op     vs base                │
_metastoreState_applyAddBlock-14   8255.60µ ± 2%   15.84µ ± 2%  -99.81% (p=0.000 n=10)
```